+++ src/main/java/com/emedicalbooking/repository/HistoryRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface HistoryRepository extends JpaRepository<History, Integer> {
+public interface HistoryRepository extends JpaRepository<History, Long> {
 
     // Lấy lịch sử khám của 1 bệnh nhân (qua booking.patient.id), sắp xếp theo thời gian tạo mới nhất
     @Query("SELECT h FROM History h " +
@@ -19,7 +19,7 @@ public interface HistoryRepository extends JpaRepository<History, Integer> {
            "LEFT JOIN FETCH b.patientProfile " +
            "WHERE b.patient.id = :patientId " +
            "ORDER BY h.createdAt DESC")
-    List<History> findByPatientId(@Param("patientId") int patientId);
+    List<History> findByPatientId(@Param("patientId") Long patientId);
 
     // Lấy tất cả lịch sử khám của 1 bác sĩ, sắp xếp theo thời gian tạo mới nhất
     @Query("SELECT h FROM History h " +
@@ -29,8 +29,8 @@ public interface HistoryRepository extends JpaRepository<History, Integer> {
            "LEFT JOIN FETCH b.timeTypeData " +
            "WHERE b.doctor.id = :doctorId " +
            "ORDER BY h.createdAt DESC")
-    List<History> findByDoctorId(@Param("doctorId") int doctorId);
+    List<History> findByDoctorId(@Param("doctorId") Long doctorId);
 
     // Kiểm tra 1 booking đã có history chưa
-    Optional<History> findByBookingId(int bookingId);
+    Optional<History> findByBookingId(Long bookingId);
 }
