+++ src/main/java/com/emedicalbooking/repository/PackageRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PackageRepository extends JpaRepository<MedicalPackage, Integer> {
+public interface PackageRepository extends JpaRepository<MedicalPackage, Long> {
 
-    List<MedicalPackage> findByClinic_Id(int clinicId);
+    List<MedicalPackage> findByClinic_Id(Long clinicId);
 }
