+++ src/main/java/com/emedicalbooking/repository/PackageServiceItemRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface PackageServiceItemRepository extends JpaRepository<PackageServiceItem, Integer> {
+public interface PackageServiceItemRepository extends JpaRepository<PackageServiceItem, Long> {
 
     @Modifying(clearAutomatically = true)
     @Query("DELETE FROM PackageServiceItem p WHERE p.medicalPackage.id = :packageId")
-    void deleteAllByPackageId(int packageId);
+    void deleteAllByPackageId(Long packageId);
 }
